refactor(plugin-ember-octane) refactor template-lint types into their own def
