+++ packages/plugin-ember-octane/src/tasks/octane-migration-status-task.ts
@@ -1,8 +1,12 @@
-import { CLIEngine, Linter } from 'eslint';
+import { CLIEngine } from 'eslint';
 import * as globby from 'globby';
 import { BaseTask, Category, Priority, Task } from '@checkup/core';
 import { OctaneMigrationStatusTaskResult } from '../results';
-import { EmberTemplateLintMessage } from '../types';
+import {
+  TemplateLintMessage,
+  TemplateLintReport,
+  TemplateLintResult,
+} from '../types/ember-template-lint';
 
 const fs = require('fs');
 const TemplateLinter = require('ember-template-lint');
@@ -90,7 +94,7 @@ export default class OctaneMigrationStatusTask extends BaseTask implements Task 
     return this.esLintEngine.executeOnFiles([`${this.rootPath}/+(app|addon)/**/*.js`]);
   }
 
-  private async runTemplateLint(): Promise<CLIEngine.LintReport> {
+  private async runTemplateLint(): Promise<TemplateLintReport> {
     let filePaths = await globby(`${this.rootPath}/+(app|addon)/**/*.hbs`);
 
     let sources = filePaths.map(path => ({
@@ -98,12 +102,12 @@ export default class OctaneMigrationStatusTask extends BaseTask implements Task 
       template: fs.readFileSync(path, { encoding: 'utf8' }),
     }));
 
-    let results: CLIEngine.LintResult[] = sources.map(({ path, template }) => {
+    let results: TemplateLintResult[] = sources.map(({ path, template }) => {
       // We need to get the ember-template-lint results and then map them over to match ESLint result
       // types. This allows us to streamline our data manipulation later.
-      let messages: Linter.LintMessage[] = this.templateLinter
+      let messages: TemplateLintMessage[] = this.templateLinter
         .verify({ source: template, moduleId: path })
-        .map((msg: EmberTemplateLintMessage) => {
+        .map((msg: TemplateLintMessage) => {
           let { column, line, message, rule, severity, source } = msg;
           return {
             column,
@@ -133,10 +137,6 @@ export default class OctaneMigrationStatusTask extends BaseTask implements Task 
     return {
       errorCount,
       results,
-      warningCount: 0,
-      fixableErrorCount: 0,
-      fixableWarningCount: 0,
-      usedDeprecatedRules: [],
     };
   }
 }
